1start/pyproject.toml
[project]
name = "start1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pyorbbecsdk-community==1.4.2",
]
